+++ examples/sequence.rs
@@ -42,7 +42,7 @@ fn main() {
 
     let wave = wave.filter(
         duration,
-        &mut (0.9 * multipass() & 0.1 * reverb_stereo(2.0)),
+        &mut (0.9 * multipass() & 0.1 * reverb_stereo(10.0, 2.0)),
     );
 
     let wave = wave.filter_latency(duration, &mut (limiter_stereo((0.05, 0.20))));